better in-process benchmark
